create-tag, reset-ui, readme: add checks
- check the actual folder in create-tag & reset-ui
- check package-lock.json in reset-ui
- remove package-lock.json mention in readme

+++ reset-ui.sh
@@ -1,11 +1,25 @@
 #!/bin/sh
 
+set -e
+
 # you have to run this script while you are in the root directory of the project
 
-echo "Removing node_modules from the root directory"
+GREEN='\033[0;32m'
+YELLOW='\033[0;33m'
+NC='\033[0m'
+
+# Check if the user is in the osrd-ui folder
+if [[ "$(pwd)" != *"/osrd-ui" ]]; then
+	echo -e "❌ ${YELLOW}You need to be in the osrd-ui folder to launch this script${NC}"
+	exit 1
+fi
+
+echo -e "${GREEN}Removing node_modules from the root directory${NC}"
 
 rm -rf node_modules
-rm package-lock.json
+if [ -f package-lock.json ]; then
+	echo -e "${YELLOW}Removing package-lock.json from the root directory${NC}"
+fi
 
 # create a list of all the osrd-ui sub-repos
 sub_repos=$(ls -d ./ui-*)
@@ -13,27 +27,29 @@ sub_repos=$(ls -d ./ui-*)
 # remove node_modules and dist from all the sub-repos
 
 echo
-echo "Removing node_modules and dist from:"
-echo $sub_repos
+echo -e "${GREEN}Removing node_modules and dist from:${NC}"
+echo -e $sub_repos
 
 for repo in $sub_repos; do
 	echo
-	echo "Removing node_modules and dist from $repo"
+	echo -e "${GREEN}Removing node_modules and dist from $repo ${NC}"
 	rm -rf $repo/node_modules
-	rm $repo/package-lock.json
+	if [ -f $repo/package-lock.json ]; then
+		echo -e "${YELLOW}Removing package-lock.json from $repo ${NC}"
+	fi
 	rm -rf $repo/dist
 done
 
 # install packages in the root directory
 
 echo
-echo "Installing packages in the root directory"
+echo -e "${GREEN}Installing packages in the root directory${NC}"
 
 npm install
 
 # build the root directory
 
 echo
-echo "Building the root directory"
+echo -e "${GREEN}Building the root directory${NC}"
 
 npm run build
